% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/height_to_mm.R
\name{height_to_mm}
\alias{height_to_mm}
\title{Pin Height to mm}
\usage{
height_to_mm(data)
}
\arguments{
\item{data}{a data frame; must have a column named either `height_cm` or `height_mm`}
}
\value{
the original data frame, but if there was a `height_cm` or `height_mm` column, it is now named `pin_height`. If original readings were in cm, they have been transformed into mm.
}
\description{
turn any pin heights into mm, in a column named pin_height
}
\examples{
df <- data.frame(site = c("SET1", "SET2"), height_cm = c(15, 18))
height_to_mm(df)

df <- data.frame(site = c("SET1", "SET2"), height_mm = c(156, 182))
height_to_mm(df)
}
