% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_change_incr.R
\name{calc_change_incr}
\alias{calc_change_incr}
\title{Incremental Change Calculations}
\usage{
calc_change_incr(dat)
}
\arguments{
\item{dat}{a data frame with one row per pin reading, and the following columns, named exactly: date, set_id, arm_position, pin_number, pin_height}
}
\value{
a list of three tibbles: one each for pin, arm, and set level calculations. Pin level change is calculated first, as the difference between a pin reading and the prior pin reading from that set_id--arm--pin. The column name in the $pin tibble is "incr". For every date of a pin reading, this calculated value will exist or be NA. On the first date, it is NA. Incremental pin changes are then averaged to the arm position level on each date, excluding NAs. St Deviation and St Error are also calculated. There is one calculated value for every arm on every reading date. These columns in the $arm tibble are "mean_incr", "sd_incr", and "se_incr". The cumulative arm changes are then averaged to the SET level, also with st dev and st err. There is one calculated value for every SET on every reading date. The columns in the $set tibble are again "mean_incr", "sd_incr", and "se_incr". Pin level calculations are the most helpful for qa/qc, as it is possible to check for and follow-up on readings that have changed more than a certain amount (e.g. 25 mm) between readings.
}
\description{
Incremental Change Calculations
}
\examples{
calc_change_incr(example_sets)
}
