% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hist_arm.R
\name{plot_hist_arm}
\alias{plot_hist_arm}
\title{Generate a histogram of pin readings by arm position}
\usage{
plot_hist_arm(data, columns = 4, scales = "free_y")
}
\arguments{
\item{data}{a data frame with one row per pin reading, and the following columns, named exactly: set_id, arm_position, pin_height}

\item{columns}{number of columns you'd like in the faceted plot}

\item{scales}{passed to `facet_wrap` - fixed or free?}
}
\value{
a ggplot object
}
\description{
Generate a histogram of pin readings by arm position
}
\examples{
plot_hist_arm(example_sets)
}
