% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_incr_arm.R
\name{plot_incr_arm}
\alias{plot_incr_arm}
\title{Plot change between readings, by arm}
\usage{
plot_incr_arm(data, threshold = 25, columns = 4, set = NULL,
  pointsize = 2, scales = "fixed")
}
\arguments{
\item{data}{data frame (e.g. `$arm` piece of output from `calc_change_incr()`) with one row per faceting variable, and the following columns, named exactly: date, set_id, arm_position, mean_incr. `mean_incr` should be an already-calculated field of change since previous reading}

\item{threshold}{numeric value for red horizontal lines (at +/- this value); this should be a value that would be a meaningful threshold for incremental change.}

\item{columns}{number of columns for faceted output}

\item{set}{optional SET ID if you only want to look at one SET; default is to graph all SETs}

\item{pointsize}{size of points you want (goes into the `size` argument of `ggplot2::geom_point`)}

\item{scales}{passed to `facet_wrap`; same fixed/free options as that function}
}
\value{
a ggplot object
}
\description{
Plot change between readings, by arm
}
\examples{
incr_set <- calc_change_incr(example_sets)
plot_incr_arm(incr_set$arm)
plot_incr_arm(incr_set$arm, threshold = 50, columns = 1)
plot_incr_arm(incr_set$arm, set = "SET2", threshold = 10)
}
