% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_incr_pin.R
\name{plot_incr_pin}
\alias{plot_incr_pin}
\title{Plot change between readings, by pin}
\usage{
plot_incr_pin(data, set, threshold = 25, columns = 2, pointsize = 2,
  scales = "fixed")
}
\arguments{
\item{data}{data frame (e.g. `$pin` piece of output from `calc_change_incr()`) with one row per faceting variable, and the following columns, named exactly: date, set_id, arm_position, pin_number, incr. `incr` should be an already-calculated field of change since previous reading.}

\item{set}{SET ID to graph (required)}

\item{threshold}{numeric value for red horizontal lines (at +/- this value); this should be a value that would be a meaningful threshold for incremental change.}

\item{columns}{number of columns for faceted output}

\item{pointsize}{size of points you want (goes into the `size` argument of `ggplot2::geom_point`)}

\item{scales}{passed to `facet_wrap`; same fixed/free options as that function}
}
\value{
a ggplot object
}
\description{
Plot change between readings, by pin
}
\examples{
incr_set <- calc_change_incr(example_sets)
plot_incr_pin(incr_set$pin, set = "SET1")
plot_incr_pin(incr_set$pin, set = "SET1", threshold = 5)
plot_incr_pin(incr_set$pin, set = "SET2", threshold = 5, columns = 1)
}
